requirements.txt
python-binance==1.0.19
aiohttp
pandas
numpy
python-dotenv
